Analyse STRUCTURE, TESS, BAPS or BASIC text runs. Wrapper around several smaller functions.

A single function to analyse STRUCTURE, TESS, BAPS or BASIC text runs. Converts runs to a qlist, tabulates, summarises, runs evanno method (for STRUCTURE runs), aligns clusters and plots all runs.

analyseQ(
  files = NULL,
  evannomethod = TRUE,
  align = TRUE,
  plotruns = TRUE,
  imgoutput = "sep",
  grplab = NA,
  clustercol = NA,
  writetable = TRUE,
  sorttable = TRUE,
  exportpath = NULL
)

Arguments

files
A character or character vector of one or more STRUCTURE, TESS, BAPS or BASIC run files. Use choose.files(multi=TRUE) to choose interactively.

evannomethod
A logical indicating if evanno method should be performed. Applies only to STRUCTURE runs.

align
A logical indicating if clusters must be aligned within and across K.

plotruns
A logical indicating if the runs should be exported as barplots.

imgoutput
A character indicating if files are plotted as separate image files ("sep") or joined into a single image ("join").

grplab
A dataframe with one or more columns (group label sets), and rows equal to the number of individuals.

clustercol
A character vector of colours for colouring clusters. If NA, colours are automatically generated. K 1 to 12 are custom unique colours while K>12 are coloured by function rich.color().

writetable
A logical TRUE or FALSE. Setting to TRUE writes the output table to the working directory.

sorttable
A logical indicating if the tabulated output table must be sorted. Sorts table by loci, ind and K when available.

exportpath
A path to export output files. To use current working directory, set exportpath=getwd().

Value

A qlist (list of data.frames) is returned.

Details

The function analyseQ is a wrapper around several other pophelper functions. All arguments for all sub-functions are not available. If more arguments/options are required, consider running the functions separately.

See the vignette for more details.

Examples

if (FALSE) {
# structure files
sfiles <- list.files(path=system.file("files/structure",package="pophelper"),
full.names=TRUE)
analyseQ(sfiles,exportpath=getwd())
}