Collect TESS cluster run files from multiple folders

Collect TESS cluster run files from multiple folders to one folder and rename each run by folder name

collectRunsTess(runsdir = NA, newdir = NA, quiet = FALSE)

Arguments

runsdir
A character path indicating the directory containing TESS runs in multiple directories. Use choose.dir() for interactively selecting the directory. If NA, or no directory is selected, the current working directory is used.

newdir
A character indicating the name of the new directory to be created with the collected runs. IF NA, the default name 'AllTESSRuns' is used.

quiet
A logical indicating if a message is to be displayed for directories without TESS runs and number of runs copied and renamed.

Value

Two integers are ruturned. The first denotes the number of TESS run files copied and renamed. The second number denotes number of directories without TESS run files.

Details

Within each TESS run folder, the function searches for filename ending with 'TR.txt' as the cluster file. This file is copied to the new folder and renamed as the name of the respective run directory. Therefore, DO NOT manually rename original run files or directories.

See the vignette for more details.

Examples

# NOT RUN {
collectRunsTess("path")
# }